Added geojson of Wahlkreise Bundestagswahl 2013
All Wahlkreise of the Bundestagswahl 2013 can be found at http://www.
bundeswahlleiter.de/de/bundestagswahlen/BTW_BUND_09/wahlkreiseinteilung/
kartogaphische_darstellung.html as a shapefile. Projection of this data is
Lambert. To get usable data we need to convert the data into WSG84 projection
and format it as geojson. We used gdal's util program ogr2ogr for this job:

  $ ogr2ogr -f GeoJSON -t_srs crs:84 wahlkreise.geojson Geometrie_...shp